pre-rel 1.0.2b
