+++ include/fix8/conjure_enum.hpp
@@ -31,6 +31,8 @@
 // OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE
 // SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 //----------------------------------------------------------------------------------------
+// conjure_enum
+//----------------------------------------------------------------------------------------
 #ifndef FIX8_CONJURE_ENUM_HPP_
 #define FIX8_CONJURE_ENUM_HPP_
 
@@ -45,12 +47,10 @@
 #include <tuple>
 #include <concepts>
 #include <optional>
-#include <functional>
-#include <ostream>
-#include <cstddef>
-#include <exception>
-#include <bitset>
-#include <bit>
+#if not defined FIX8_CONJURE_ENUM_MINIMAL
+# include <functional>
+#endif
+//#include <cstddef>
 #include <array>
 
 //-----------------------------------------------------------------------------------------
@@ -174,7 +174,8 @@ template<valid_enum T>
 class conjure_enum : public static_only
 {
 	static constexpr int enum_min_value{enum_range<T>::min}, enum_max_value{enum_range<T>::max};
-	static_assert(enum_max_value > enum_min_value, "FIX8_CONJURE_ENUM_MAX_VALUE must be greater than FIX8_CONJURE_ENUM_MIN_VALUE");
+	static_assert(enum_max_value > enum_min_value,
+		"FIX8_CONJURE_ENUM_MAX_VALUE (or enum_range<T>::max) must be greater than FIX8_CONJURE_ENUM_MIN_VALUE (or enum_range<T>::min) ");
 
 public:
 	using enum_tuple = std::tuple<T, std::string_view>;
@@ -197,57 +198,12 @@ private:
 		return std::array<enum_tuple, sizeof...(I)>{{{ values[I], _enum_name_v<values[I]>}...}};
 	}
 
-	template<std::size_t... I>
-	static constexpr auto _names(std::index_sequence<I...>) noexcept
-	{
-		return std::array<std::string_view, sizeof...(I)>{{{ _enum_name_v<values[I]>}...}};
-	}
-
 	static constexpr auto _sorted_entries() noexcept
 	{
 		auto tmp { entries };
 		std::sort(tmp.begin(), tmp.end(), _tuple_comp_rev);
 		return tmp;
 	}
-
-#if not defined FIX8_CONJURE_ENUM_MINIMAL
-	template<std::size_t... I>
-	static constexpr auto _unscoped_entries(std::index_sequence<I...>) noexcept
-	{
-		std::array<enum_tuple, sizeof...(I)> tmp{{{ values[I], _remove_scope(_enum_name_v<values[I]>)}...}};
-		std::sort(tmp.begin(), tmp.end(), _tuple_comp_rev);
-		return tmp;
-	}
-
-	static constexpr std::string_view _remove_scope(std::string_view what) noexcept
-	{
-		if (const auto lc { what.find_last_of(':') }; lc != std::string_view::npos)
-			return what.substr(lc + 1);
-		return what;
-	}
-
-	template<std::size_t... I>
-	static constexpr auto _scoped_entries(std::index_sequence<I...>) noexcept
-	{
-		std::array<scoped_tuple, sizeof...(I)> tmp{{{ _remove_scope(_enum_name_v<values[I]>), _enum_name_v<values[I]>}...}};
-		std::sort(tmp.begin(), tmp.end(), _scoped_comp);
-		return tmp;
-	}
-
-	template<std::size_t... I>
-	static constexpr auto _rev_scoped_entries(std::index_sequence<I...>) noexcept
-	{
-		std::array<scoped_tuple, sizeof...(I)> tmp{{{ _enum_name_v<values[I]>, _remove_scope(_enum_name_v<values[I]>)}...}};
-		std::sort(tmp.begin(), tmp.end(), _scoped_comp);
-		return tmp;
-	}
-
-	template<std::size_t... I>
-	static constexpr auto _unscoped_names(std::index_sequence<I...>) noexcept
-	{
-		return std::array<std::string_view, sizeof...(I)>{{{ _remove_scope(_enum_name_v<values[I]>)}...}};
-	}
-#endif
 
 	template<T e>
 	static constexpr bool _is_valid() noexcept
@@ -314,18 +270,6 @@ private:
 			return {};
 	}
 
-#if not defined FIX8_CONJURE_ENUM_MINIMAL
-	static constexpr std::string_view _process_scope([[maybe_unused]] const auto& entr, std::string_view what) noexcept
-	{
-		if constexpr (is_scoped())
-			if (const auto result { std::equal_range(entr.cbegin(),
-				entr.cend(), scoped_tuple(what, std::string_view()), _scoped_comp) };
-					result.first != result.second)
-						return std::get<1>(*result.first);
-		return what;
-	}
-#endif
-
 	/// comparators
 	static constexpr bool _value_comp(const T& pl, const T& pr) noexcept
 	{
@@ -339,12 +283,6 @@ private:
 	{
 		return std::get<std::string_view>(pl) < std::get<std::string_view>(pr);
 	}
-#if not defined FIX8_CONJURE_ENUM_MINIMAL
-	static constexpr bool _scoped_comp(const scoped_tuple& pl, const scoped_tuple& pr) noexcept
-	{
-		return std::get<0>(pl) < std::get<0>(pr);
-	}
-#endif
 
 public:
 	static consteval const char *tpeek() noexcept { return std::source_location::current().function_name(); }
@@ -397,25 +335,6 @@ public:
 		else
 			return false;
 	}
-#if not defined FIX8_CONJURE_ENUM_MINIMAL
-	static constexpr std::string_view remove_scope(std::string_view what) noexcept
-	{
-		return _process_scope(rev_scoped_entries, what);
-	}
-
-	static constexpr std::string_view add_scope(std::string_view what) noexcept
-	{
-		return _process_scope(scoped_entries, what);
-	}
-#endif
-
-	// iterators
-	static constexpr auto cbegin() noexcept { return entries.cbegin(); }
-	static constexpr auto cend() noexcept { return entries.cend(); }
-	static constexpr auto crbegin() noexcept { return entries.crbegin(); }
-	static constexpr auto crend() noexcept { return entries.crend(); }
-	static constexpr auto front() noexcept { return *cbegin(); }
-	static constexpr auto back() noexcept { return *std::prev(cend()); }
 
 	// enum <==> int
 	static constexpr int enum_to_int(T value) noexcept
@@ -483,452 +402,21 @@ public:
 				return std::get<T>(*result.first);
 		return {};
 	}
-#if not defined FIX8_CONJURE_ENUM_MINIMAL
-	static constexpr std::optional<T> unscoped_string_to_enum(std::string_view str) noexcept
-	{
-		if (const auto result { std::equal_range(unscoped_entries.cbegin(), unscoped_entries.cend(), enum_tuple(T{}, str), _tuple_comp_rev) };
-			result.first != result.second)
-				return std::get<T>(*result.first);
-		return {};
-	}
-#endif
-
-	/// for_each, for_each_n
-	template<typename Fn, typename... Args>
-	requires std::invocable<Fn&&, T, Args...>
-	[[maybe_unused]] static constexpr auto for_each(Fn&& func, Args&&... args) noexcept
-	{
-		return for_each_n(static_cast<int>(count()), std::forward<Fn>(func), std::forward<Args>(args)...);
-	}
-
-	template<typename Fn, typename C, typename... Args> // specialisation for member function with object
-	requires std::invocable<Fn&&, C, T, Args...>
-	[[maybe_unused]] static constexpr auto for_each(Fn&& func, C *obj, Args&&... args) noexcept
-	{
-		return for_each(std::bind(std::forward<Fn>(func), obj, std::placeholders::_1, std::forward<Args>(args)...));
-	}
-
-	template<typename Fn, typename... Args>
-	requires std::invocable<Fn&&, T, Args...>
-	[[maybe_unused]] static constexpr auto for_each_n(int n, Fn&& func, Args&&... args) noexcept
-	{
-		for (int ii{}; const auto ev : conjure_enum<T>::values)
-		{
-			if (ii++ >= n)
-				break;
-			std::invoke(std::forward<Fn>(func), ev, std::forward<Args>(args)...);
-		}
-		return std::bind(std::forward<Fn>(func), std::placeholders::_1, std::forward<Args>(args)...);
-	}
-
-	template<typename Fn, typename C, typename... Args> // specialisation for member function with object
-	requires std::invocable<Fn&&, C, T, Args...>
-	[[maybe_unused]] static constexpr auto for_each_n(int n, Fn&& func, C *obj, Args&&... args) noexcept
-	{
-		return for_each_n(n, std::bind(std::forward<Fn>(func), obj, std::placeholders::_1, std::forward<Args>(args)...));
-	}
-
-	// dispatch
-	template<typename Fn>
-	static constexpr bool tuple_comp(const std::tuple<T, Fn>& pl, const std::tuple<T, Fn>& pr) noexcept
-	{
-		return static_cast<int>(std::get<T>(pl)) < static_cast<int>(std::get<T>(pr));
-	}
-
-	template<std::size_t I, typename R, typename Fn, typename... Args> // with not found value(nval) for return
-	requires std::invocable<Fn&&, T, Args...>
-	[[maybe_unused]] static constexpr R dispatch(T ev, R nval, const std::array<std::tuple<T, Fn>, I>& disp, Args&&... args) noexcept
-	{
-		const auto result { std::equal_range(disp.cbegin(), disp.cend(), std::make_tuple(ev, Fn()), tuple_comp<Fn>) };
-		return result.first != result.second ? std::invoke(std::get<Fn>(*result.first), ev, std::forward<Args>(args)...) : nval;
-	}
-
-	template<std::size_t I, typename R, typename Fn, typename C, typename... Args> // specialisation for member function with not found value(nval) for return
-	requires std::invocable<Fn&&, C, T, Args...>
-	[[maybe_unused]] static constexpr R dispatch(T ev, R nval, const std::array<std::tuple<T, Fn>, I>& disp, C *obj, Args&&... args) noexcept
-	{
-		const auto result { std::equal_range(disp.cbegin(), disp.cend(), std::make_tuple(ev, Fn()), tuple_comp<Fn>) };
-		return result.first != result.second ? std::invoke(std::get<Fn>(*result.first), obj, ev, std::forward<Args>(args)...) : nval;
-	}
-
-	template<std::size_t I, typename Fn, typename... Args> // void func with not found call to last element
-	requires (std::invocable<Fn&&, T, Args...> && I > 0)
-	static constexpr void dispatch(T ev, const std::array<std::tuple<T, Fn>, I>& disp, Args&&... args) noexcept
-	{
-		const auto result { std::equal_range(disp.cbegin(), std::prev(disp.cend()), std::make_tuple(ev, Fn()), tuple_comp<Fn>) };
-		return result.first != result.second ? std::invoke(std::get<Fn>(*result.first), ev, std::forward<Args>(args)...)
-														 : std::invoke(std::get<Fn>(*std::prev(disp.cend())), ev, std::forward<Args>(args)...);
-	}
-
-	template<std::size_t I, typename Fn, typename C, typename... Args> // specialisation for void member function with not found call to last element
-	requires (std::invocable<Fn&&, C, T, Args...> && I > 0)
-	static constexpr void dispatch(T ev, const std::array<std::tuple<T, Fn>, I>& disp, C *obj, Args&&... args) noexcept
-	{
-		const auto result { std::equal_range(disp.cbegin(), std::prev(disp.cend()), std::make_tuple(ev, Fn()), tuple_comp<Fn>) };
-		return result.first != result.second ? std::invoke(std::get<Fn>(*result.first), obj, ev, std::forward<Args>(args)...)
-														 : std::invoke(std::get<Fn>(*std::prev(disp.cend())), obj, ev, std::forward<Args>(args)...);
-	}
 
 	// public constexpr data structures
 	static constexpr auto values { _values(std::make_index_sequence<enum_max_value - enum_min_value + 1>()) };
 	static constexpr auto entries { _entries(std::make_index_sequence<values.size()>()) };
-	static constexpr auto names { _names(std::make_index_sequence<values.size()>()) };
 	static constexpr auto sorted_entries { _sorted_entries() };
-#if not defined FIX8_CONJURE_ENUM_MINIMAL
-	static constexpr auto scoped_entries { _scoped_entries(std::make_index_sequence<values.size()>()) };
-	static constexpr auto unscoped_entries { _unscoped_entries(std::make_index_sequence<values.size()>()) };
-	static constexpr auto rev_scoped_entries { _rev_scoped_entries(std::make_index_sequence<values.size()>()) };
-	static constexpr auto unscoped_names { _unscoped_names(std::make_index_sequence<values.size()>()) };
-#endif
 
 	// misc
 	static constexpr int get_enum_min_value() noexcept { return enum_min_value; }
 	static constexpr int get_enum_max_value() noexcept { return enum_max_value; }
-};
 
-//-----------------------------------------------------------------------------------------
-// allow range based for
-template<valid_enum T>
-struct iterator_adaptor
-{
-	constexpr auto begin() noexcept { return conjure_enum<T>::entries.cbegin(); }
-	constexpr auto end() noexcept { return conjure_enum<T>::entries.cend(); }
-};
-
-//-----------------------------------------------------------------------------------------
-// ostream& operator<< for any enum; add the following before using:
-// using ostream_enum_operator::operator<<;
-//-----------------------------------------------------------------------------------------
-namespace ostream_enum_operator
-{
-	template<typename CharT, typename Traits=std::char_traits<CharT>, valid_enum T>
-	constexpr std::basic_ostream<CharT, Traits>& operator<<(std::basic_ostream<CharT, Traits>& os, T value) noexcept
-	{
-		if (conjure_enum<T>::contains(value))
-			return os << conjure_enum<T>::enum_to_string(value);
-		return os << conjure_enum<T>::enum_to_underlying(value);
-	}
-}
-
-//-----------------------------------------------------------------------------------------
-//-----------------------------------------------------------------------------------------
-template<typename T>
-concept valid_bitset_enum = valid_enum<T> and requires(T)
-{
-	requires static_cast<std::size_t>(conjure_enum<T>::values.back()) < conjure_enum<T>::count();
-};
-
-//-----------------------------------------------------------------------------------------
-// bitset based on supplied enum
-// Note: your enum sequence must be continuous with the last enum value < count of enumerations
-//-----------------------------------------------------------------------------------------
-template<valid_bitset_enum T>
-class enum_bitset
-{
-	using U = std::underlying_type_t<T>;
-	static constexpr auto countof { conjure_enum<T>::count() };
-
-	template<T val>
-	static constexpr U to_underlying() noexcept { return static_cast<U>(val); } // C++23: upgrade to std::to_underlying
-	static constexpr U to_underlying(T val) noexcept { return static_cast<U>(val); }
-	static constexpr U all_bits { (1 << countof) - 1 };
-	U _present{};
-
-public:
-	explicit constexpr enum_bitset(U bits) noexcept : _present(bits) {}
-	constexpr enum_bitset(std::string_view from, bool anyscope=false, char sep='|', bool ignore_errors=true)
-		: _present(factory(from, anyscope, sep, ignore_errors)) {}
-
-	template<valid_bitset_enum... E>
-	constexpr enum_bitset(E... comp) noexcept : _present((0u | ... | (1 << to_underlying(comp)))) {}
-
-	template<std::integral... I>
-	constexpr enum_bitset(I... comp) noexcept : _present((0u | ... | (1 << comp))) {}
-
-	constexpr enum_bitset() = default;
-	constexpr ~enum_bitset() = default;
-
-	constexpr std::size_t count() const noexcept
-		{ return std::popcount(static_cast<std::make_unsigned_t<U>>(_present)); } // C++23: upgrade to std::bitset when count is constexpr
-	constexpr std::size_t not_count() const noexcept { return countof - count(); }
-	constexpr std::size_t size() const noexcept { return countof; }
-	constexpr U to_ulong() const noexcept { return _present; }
-	constexpr unsigned long long to_ullong() const noexcept { return _present; }
-
-	constexpr bool operator[](std::size_t pos) const noexcept { return test(pos); }
-	constexpr bool operator[](T what) const noexcept { return test(what); }
-
-	/// set
-	constexpr void set(U pos, bool value=true) noexcept { value ? _present |= 1 << pos : _present &= ~(1 << pos); }
-	constexpr void set(T what, bool value=true) noexcept { set(to_underlying(what), value); }
-	constexpr void set() noexcept { _present = all_bits; }
-
-	template<T what>
-	constexpr void set() noexcept
-	{
-		if constexpr (constexpr auto uu{to_underlying<what>()}; uu < countof)
-			_present |= 1 << uu;
-	}
-
-	template<T... comp>
-	requires (sizeof...(comp) > 1)
-	constexpr void set() noexcept { (set<comp>(),...); }
-
-	template<valid_bitset_enum... E>
-	requires (sizeof...(E) > 1)
-	constexpr void set(E... comp) noexcept { return (... | (set(comp))); }
-
-	/// flip
-	template<T what>
-	constexpr void flip() noexcept
-	{
-		if constexpr (constexpr auto uu{to_underlying<what>()}; uu < countof)
-			_present ^= 1 << uu;
-	}
-
-	constexpr void flip() noexcept { _present = ~_present & all_bits; }
-	constexpr void flip(U pos) noexcept { _present ^= 1 << pos; }
-	constexpr void flip(T what) noexcept { flip(to_underlying(what)); }
-
-	/// reset
-	template<T what>
-	constexpr void reset() noexcept
-	{
-		if constexpr (constexpr auto uu{to_underlying<what>()}; uu < countof)
-			_present &= ~(1 << uu);
-	}
-
-	constexpr void reset() noexcept { _present = 0; }
-	constexpr void reset(U pos) noexcept { _present &= ~(1 << pos); }
-	constexpr void reset(T what) noexcept { reset(to_underlying(what)); }
-
-	template<T... comp>
-	requires (sizeof...(comp) > 1)
-	constexpr void reset() noexcept { (reset<comp>(),...); }
-
-	template<std::integral... I>
-	constexpr void reset(I...comp) noexcept { (reset(comp),...); }
-
-	/// test
-	constexpr bool test(U pos) const noexcept { return _present & 1 << pos; }
-	constexpr bool test(T what) const noexcept { return test(to_underlying(what)); }
-	constexpr bool test() const noexcept { return _present; }
-
-	template<T what>
-	constexpr bool test() const noexcept
-	{
-		if constexpr (constexpr auto uu{to_underlying<what>()}; uu < countof)
-			return test(uu);
-		else
-			return false;
-	}
-
-	template<T... comp>
-	constexpr bool any_of() const noexcept { return (... || test<comp>()); }
-
-	template<std::integral... I>
-	constexpr bool any_of(I...comp) const noexcept { return (... || test(comp)); }
-
-	template<valid_bitset_enum... E>
-	constexpr bool any_of(E...comp) const noexcept { return (... || test(comp)); }
-
-	template<T... comp>
-	constexpr bool all_of() const noexcept { return (... && test<comp>()); }
-
-	template<std::integral... I>
-	constexpr bool all_of(I...comp) const noexcept { return (... && test(comp)); }
-
-	template<valid_bitset_enum... E>
-	constexpr bool all_of(E...comp) const noexcept { return (... && test(comp)); }
-
-	template<T... comp>
-	constexpr bool none_of() const noexcept { return (... && !test<comp>()); }
-
-	template<std::integral... I>
-	constexpr bool none_of(I...comp) const noexcept { return (... && !test(comp)); }
-
-	template<valid_bitset_enum... E>
-	constexpr bool none_of(E...comp) const noexcept { return (... && !test(comp)); }
-
-	constexpr bool any() const noexcept { return count(); }
-	constexpr bool all() const noexcept { return _present == all_bits; }
-	constexpr bool none() const noexcept { return !*this; }
-
-	/// operators
-	constexpr operator bool() const noexcept { return count(); }
-	constexpr enum_bitset& operator<<=(std::size_t pos) noexcept { _present <<= pos; return *this; }
-	constexpr enum_bitset& operator>>=(std::size_t pos) noexcept { _present >>= pos; return *this; }
-	constexpr enum_bitset& operator&=(T other) noexcept { _present &= 1 << to_underlying(other); return *this; }
-	constexpr enum_bitset& operator|=(T other) noexcept { _present |= 1 << to_underlying(other); return *this; }
-	constexpr enum_bitset& operator^=(T other) noexcept { _present ^= 1 << to_underlying(other); return *this; }
-	constexpr enum_bitset& operator&=(U other) noexcept { _present &= other; return *this; }
-	constexpr enum_bitset& operator|=(U other) noexcept { _present |= other; return *this; }
-	constexpr enum_bitset& operator^=(U other) noexcept { _present ^= other; return *this; }
-
-	constexpr enum_bitset operator<<(int pos) const noexcept { return enum_bitset(_present << pos); }
-	constexpr enum_bitset operator>>(int pos) const noexcept { return enum_bitset(_present >> pos); }
-	constexpr enum_bitset operator&(T other) const noexcept { return enum_bitset(_present & 1 << to_underlying(other)); }
-	constexpr enum_bitset operator|(T other) const noexcept { return enum_bitset(_present | 1 << to_underlying(other)); }
-	constexpr enum_bitset operator^(T other) const noexcept { return enum_bitset(_present ^ 1 << to_underlying(other)); }
-	constexpr enum_bitset operator&(U other) const noexcept { return enum_bitset(_present & other); }
-	constexpr enum_bitset operator|(U other) const noexcept { return enum_bitset(_present | other); }
-	constexpr enum_bitset operator^(U other) const noexcept { return enum_bitset(_present ^ other); }
-	constexpr enum_bitset operator~() const noexcept { return enum_bitset(~_present & all_bits); }
-
-	/// for_each, for_each_n
-	template<typename Fn, typename... Args>
-	requires std::invocable<Fn&&, T, Args...>
-	[[maybe_unused]] constexpr auto for_each(Fn&& func, Args&&... args) noexcept
-	{
-		return for_each_n(static_cast<int>(countof), std::forward<Fn>(func), std::forward<Args>(args)...);
-	}
-
-	template<typename C, typename Fn, typename... Args> // specialisation for member function with object
-	requires std::invocable<Fn&&, C, T, Args...>
-	[[maybe_unused]] constexpr auto for_each(Fn&& func, C *obj, Args&&... args) noexcept
-	{
-		return for_each(std::bind(std::forward<Fn>(func), obj, std::placeholders::_1, std::forward<Args>(args)...));
-	}
-
-	template<typename Fn, typename... Args>
-	requires std::invocable<Fn&&, T, Args...>
-	[[maybe_unused]] constexpr auto for_each_n(int n, Fn&& func, Args&&... args) noexcept
-	{
-		for (int ii{}, jj{}; ii < static_cast<int>(countof) && jj < n; ++ii)
-			if (const auto ev{conjure_enum<T>::values[ii]}; test(ev))
-				std::invoke(std::forward<Fn>(func), ev, std::forward<Args>(args)...), ++jj;
-		return std::bind(std::forward<Fn>(func), std::placeholders::_1, std::forward<Args>(args)...);
-	}
-
-	template<typename C, typename Fn, typename... Args> // specialisation for member function with object
-	requires std::invocable<Fn&&, C, T, Args...>
-	[[maybe_unused]] constexpr auto for_each_n(int n, Fn&& func, C *obj, Args&&... args) noexcept
-	{
-		return for_each_n(n, std::bind(std::forward<Fn>(func), obj, std::placeholders::_1, std::forward<Args>(args)...));
-	}
-
-	/// create a bitset from custom separated enum string
-	static constexpr U factory(std::string_view src, bool anyscope, char sep, bool ignore_errors)
-	{
-		enum_bitset result;
-		constexpr auto trim([](std::string_view src) noexcept ->auto
-		{
-		   const auto bg(src.find_first_not_of(" \t"));
-			return bg == std::string_view::npos ? src : src.substr(bg, src.find_last_not_of(" \t") - bg + 1);
-		});
-		auto process([anyscope,&result](std::string_view src) noexcept ->auto
-		{
-			if (anyscope && !conjure_enum<T>::has_scope(src))
-				src = conjure_enum<T>::add_scope(src);
-			if (auto ev { conjure_enum<T>::string_to_enum(src) }; ev)
-			{
-				result |= *ev;
-				return true;
-			}
-			return false;
-		});
-		for (std::string_view::size_type pos{}, fnd{};; pos = fnd + 1)
-		{
-			if ((fnd = src.find_first_of(sep, pos)) != std::string_view::npos)
-			{
-				if (auto srcp { trim(src.substr(pos, fnd - pos)) }; !process(trim(srcp)) && !ignore_errors)
-					throw std::invalid_argument(std::string(srcp).c_str());
-				continue;
-			}
-			if (pos < src.size())
-				if (auto srcp { trim(src.substr(pos, src.size() - pos)) }; !process(trim(srcp)) && !ignore_errors)
-					throw std::invalid_argument(std::string(srcp).c_str());
-			break;
-		}
-      return result._present;
-	}
-
-	constexpr std::string to_string(char zero='0', char one='1') const noexcept
-	{
-		return std::bitset<countof>(_present).to_string(zero, one);
-	}
-
-	friend constexpr std::ostream& operator<<(std::ostream& os, const enum_bitset& what) noexcept
-	{
-		return os << what.to_string();
-	}
-};
-
-template<typename T>
-constexpr enum_bitset<T> operator&(const enum_bitset<T>& lh, const enum_bitset<T>& rh) noexcept
-	{ return lh.operator&(rh.to_ulong()); }
-template<typename T>
-constexpr enum_bitset<T> operator|(const enum_bitset<T>& lh, const enum_bitset<T>& rh) noexcept
-	{ return lh.operator|(rh.to_ulong()); }
-template<typename T>
-constexpr enum_bitset<T> operator^(const enum_bitset<T>& lh, const enum_bitset<T>& rh) noexcept
-	{ return lh.operator^(rh.to_ulong()); }
-
-//-----------------------------------------------------------------------------------------
-// General purpose class allowing you to extract a string representation of any typename.
-// The string will be stored statically by the compiler, so you can use the statically generated value `name` to obtain your type.
-//-----------------------------------------------------------------------------------------
-template<typename T>
-class conjure_type : public static_only
-{
-	static constexpr std::string_view _get_name() noexcept
-	{
-		constexpr std::string_view from{tpeek()};
-#if defined _MSC_VER
-		constexpr auto ep { from.rfind(cs::get_spec<sval::start,stype::type_t>()) };
-		if constexpr (ep == std::string_view::npos)
-			return {};
-		if constexpr (constexpr auto lc { from.find_first_of(cs::get_spec<sval::end,stype::type_t>()) }; lc != std::string_view::npos)
-		{
-			constexpr auto e1 { from.substr(lc + 1, ep - lc - 2) };
-			CHKMSSTR(e1,type_t);
-			CHKMSSTR(e1,extype_t0);
-			CHKMSSTR(e1,extype_t1);
-			CHKMSSTR(e1,extype_t2);
-			CHKMSSTR(e1,extype_t3);
-		}
-		return {};
-	}
+#if not defined FIX8_CONJURE_ENUM_MINIMAL
+#include <fix8/conjure_enum_ext.hpp>
 #else
-		constexpr auto ep { from.rfind(cs::get_spec<sval::start,stype::type_t>()) };
-		if constexpr (ep == std::string_view::npos)
-			return {};
-		if constexpr (from[ep + cs::get_spec<sval::start,stype::type_t>().size()] == cs::get_spec<sval::anon_start,stype::type_t>())
-			if (constexpr auto lstr { from.substr(ep + cs::get_spec<sval::start,stype::type_t>().size()) };
-				lstr.find(cs::get_spec<sval::anon_str,stype::type_t>()) != std::string_view::npos)	// is anon
-					if constexpr (constexpr auto lc { lstr.find_first_of(cs::get_spec<sval::end,stype::type_t>()) }; lc != std::string_view::npos)
-						return lstr.substr(cs::get_spec<sval::anon_str,stype::type_t>().size() + 2, lc - (cs::get_spec<sval::anon_str,stype::type_t>().size() + 2)); // eat "::"
-		constexpr auto result { from.substr(ep + cs::get_spec<sval::start,stype::type_t>().size()) };
-		if constexpr (constexpr auto lc { result.find_first_of(cs::get_spec<sval::end,stype::type_t>()) }; lc != std::string_view::npos)
-			return result.substr(0, lc);
-		return {};
-	}
-	static constexpr auto _type_name() noexcept
-	{
-		constexpr auto result { _get_name() };
-		return fixed_string<result.size()>(result);
-	}
-#endif
-
-public:
-	static consteval const char *tpeek() noexcept { return std::source_location::current().function_name(); }
-	static constexpr auto name
-	{
-#if defined _MSC_VER
-		_get_name()
-#else
-		_type_name()
-#endif
-	};
-	static constexpr std::string_view as_string_view() noexcept
-	{
-#if defined _MSC_VER
-		return name;
-#else
-		return name.get();
-#endif
-	}
 };
+#endif
 
 //-----------------------------------------------------------------------------------------
 } // FIX8
